CB-22286 Improve Azure and AWS attached disk handling at upscale (pod restart)
Problem:

During the upscale process, scenarios can occur where disks remain unattached in both Azure and AWS environments. This issue leads to inconsistencies and inefficiencies in disk management. Specifically, the following problems have been identified:

Azure example:

1) Virtual Machine (VM) provisioning is completed.
2) Disk creation takes place.
3) The Cloudbreak POD is terminated, and the flow is restarted.
4) VM provisioning is not repeated (which is desired behavior).
5) Disk creation occurs again (which is undesirable).
6) Another disk resource is created.

A similar situation can arise during POD restart when disk attachments are made, but only some of the disks are successfully attached. For instance, if 10 disks per VM are requested, certain nodes may end up with 14 disks because during the initial flow run, 4 disks were attached.

Another issue in Azure is the sequential attachment process, where attachments occur one by one for each VM. This behavior is suboptimal and requires improvement.

AWS: Similar issues are observed in AWS, except the disk attachment process occurs in parallel.

Fixes implemented in this PR:

The following scenarios were tested and fixed for both Azure and AWS environments:

1) During instance upscale, the volume is created and saved in the database, JVM stopped. After a Cloudbreak restart, the volume attachment is completed and saved correctly.

2) During instance upscale, the volume is created but not saved in the database because the JVM stopped. After a Cloudbreak restart, the volume attachment is performed successfully and saved appropriately.

3) During instance upscale, the volume is created and saved in the database. The attachment is completed but not saved, JVM stopped. After a Cloudbreak restart, the volume attachment is executed correctly and saved as intended.

4) During instance upscale, the volume is created and saved in the database. The attachment is successfully performed and saved, but JVM stopped. After a Cloudbreak restart, all operations are executed without issues.

To improve efficiency, all disk attachment (and detachment) operations for a VM now occur in a single API call, reducing overhead and enhancing the overall disk handling process.

Another bug was identified and resolved as follows: When attempting to repair a node and the attachment process failed, it was discovered that the volumeset was mistakenly deleted during the rollback phase. This deletion was deemed undesirable and has been rectified.