CB-22367 FreeIPA Service Downscale must support AZ configurations
For Multi Az stack on Azure, following needs to be done as part of downscale :

Validate that downscale does not result in number of nodes less than needed for multi Az deployment. Downscaling can be done based on explicitly specifying nodes to delete or target availability type. It should be validated in both the cases

Nodes to be removed should be chosen in a way that nodes remain uniformly distributed across availability zones after downscale. For Ex. If we have MultiAZ FreeIP with 3 nodes with 2 nodes in AZ 1 and 1 node in AZ2 and we want to downscale it to 2 nodes, node from AZ1 should be removed