CB-21702 Rotate FreeIpa admin password
The goal of this commit is to enable rotation of admin password. As the directory password is the same as admin password we rotate them at the same time to the same value. As it is more complex to rotate directory manager password pre RHEL8, we enable rotation of freeipa admin password and directory manager password on RHEL8 only. Salt pillars are also updated if the admin password and directory password was rotated successfully. Rollback is also implemented in case of failure.

This commit does not contain secrets