CB-22069 Instance types supported in all availability zones for instance group should only be allowed for vertical scaling for multi Az cluster
While fetching instance type for vertical scaling, CB should accept Availability zones as query parameters which will be used as additional filter for instance types. If filter for availability zones is provided in request then CB should return only those instance type which are supported only in availability zones provided as filter.

For GCP and AWS, filtering based on Availability zone is not being done so we are not planning to change it as part of this. Filtering for availability zones should be applicable for Azure only.