CB-22566. Switch IPA backup to use default S3 encryption.
An explicit sse flag is no longer required when attempting to upload a
backup, since S3 introduced default encryption. If the bucket is
configured with a default KMS key - that will be used.

Testing: Modified the script on a running IPA node, and verified the
encryption being set to the bucket level KMS key for IPA backups.