CB-22649 NATIVE AWS & GCP - Disks are deleted at repair if one upscale fails
Scenario:

AWS native, GCP (others were fixed previously)

Repair 3 nodes. 2 of them upscaled successfully, but one of them failed. The rollback will run for every node, and we will delete the disks because after the successful disk attachment, the instance ids are updated, and the delete on termination flag is set to true (it is set to true at the downscale part of the repair)