CB-22524 Implement fallback logic from Marketplace to VHD images for upscale flows
Extended core and freeipa upscale flows with a new flow step: IMAGE_FALLBACK
Whenever a ManagementException, with the code 'MarketplacePurchaseEligibilityFailed' is thrown
in the ADD_INSTANCES flow step, the flow proceeds to the IMAGE_FALLBACK flow step, where the logic
attempts to find a vhd image for the same image id instead of the Marketplace image that was
used to attempt adding the new isntances in the first place.
If a vhd image is found, then the flow proceeds to the ADD_INSTANCES step again to attempt to
add the new instances using the vhd image.