CB-21976 do a secondary environment validation just prior to saving the sdx cluster into the database.
we make API calls at sdx creation phase, which can take time. Moving the heavier calls to flow instead of rest would be better, but it's a longer task. There is an edge case where env delete is called right after env validation in datalake, but the datalake object isn't saved in the database. Environment calls the datalake service to fetch if any datalake is attached to the env, and datalake will return that there is no datalake for the env. At the same time, the datalake rest call is processed by datalake and it saves the datalake to the database. In this scenario, datalake will be dangling without environment because environment delete proceeds. The aim of this implementation is to reduce the duration of this specific window scenario.
=