CB-22824: Delete the Spark2 cluster templates in 7.2.18
This is a follow up task for the Spark2 deprecation.
The Spark2 7.2.18 templates will be deleteted,
Spark2 will be still available in the CDH parcel
but it can be only installed via custom templates
if a customer would like to do so.

These Spark2 templates will be deleted:

7.2.18 - Data Engineering: Apache Spark, Apache Hive, Apache Oozie
7.2.18 - Data Engineering: HA: Apache Spark, Apache Hive, Apache Oozie
7.2.18 - Real-time Data Mart: Apache Impala, Hue, Apache Kudu, Apache Spark
7.2.18 - Data Discovery and Exploration

New template:

7.2.18 - Data Discovery and Exploration for Spark3

Other changes which are necessary to make it in sync with the Spark2 blueprints:

- "7.2.18 - Data Engineering: Apache Spark3" blueprint was
  renamed to "7.2.18 - Data Engineering: Apache Spark3, Apache Hive, Apache Oozie"
- The QUERY_PROCESSOR service has been added to the
  7.2.16, 7.2.17, 7.2.18 blueprints: cdp-data-engineering-spark3-ha.bp,
  cdp-data-engineering-spark3.bp because it was part of
  the Spark2 blueprints: cdp-data-engineering.bp, cdp-data-engineering-ha.bp

These are the Spark3 versions which are equivalent
with the above deleted Spark2 templates:

7.2.18 - Data Engineering: HA: Apache Spark3, Apache Hive, Apache Oozie
7.2.18 - Data Engineering: Apache Spark3, Apache Hive, Apache Oozie
7.2.18 - Real-time Data Mart: Apache Impala, Hue, Apache Kudu, Apache Spark3
7.2.18 - Data Discovery and Exploration for Spark3