CB-22777: Add TargetMajorVersion enum elements without underscore
Due to a swagger-codegen library issue, the values of the TargetMajorVersion enum are not properly generated into Java code.
The problem is that due to an underscore being present in the elements, the swagger-codegen library converted the `VERSION_11` into `_11`
So I've added the elements without the underscore and we'll remove the old elements in the next release to avoid compatibility issues.

GitHub issue about the swagger-codegen issue: https://github.com/swagger-api/swagger-codegen/issues/10876