CB-22472 Azure Marketplace Image Upgrade e2e tests
Added upgrade tests for FreeIPA and SDX, where the nodes
are created with VHD images and during the upgrade they
are replaced with Azure MArketplace images.

Fixed a bug, where an enum was campared with a string,
to determine wether the marketplace image terms policy
setting should be queried from the environment service.