CB-22983 Remove DNS zone test from FreeIPA upgrade test
When a replica is installed it tries to add itself to all DNS zone as a nameserver. If we remove this DNS zone during replica installation it would fail:
```
2023-08-26T02:37:52Z DEBUG The ipa-replica-install command failed, exception: NotFound: 1.0.10.in-addr.arpa.: DNS zone not found
2023-08-26T02:37:52Z ERROR 1.0.10.in-addr.arpa.: DNS zone not found
2023-08-26T02:37:52Z ERROR The ipa-replica-install command failed. See /var/log/ipareplica-install.log for more information
```
As it is a very rare scenario that a DNS zone is added or removed we can ignore this issue for now.