CB-23040 Enable Azure Flexible Server to existing environments aka Env edit
There is a new field in the network request called flexibleServerSubnetIds. For existing environments customers need to add a delegated subnet so the Data hubs can start with a Flexible Server.
Additionally to the subnet ids we need the private DNS zone as well because they are both needed for the private setup.

 Flexible server subnet id modification validation:
* In case of existing network setup there must be minimum one flexible server delegated network
* you can remove the already used network, the new clusters will be started with one of the newly added subnet
* delegation type validation
* validate that all the subnets are existing on the provider side
DNS zone validation:
* currently validation is only happening during environment creation. If new DNS zone is set during env edit, we have to validate in the same way as during creation.
* existing dns zone removal is not possible, only change is enabled