CB-23089 Make separate threadpool for meteringv2 quartz job
There is no way to specify different threadpools on job level in Quartz, we have to create a separate scheduler for metering. This is required to separate sync and metering related job executions and prevent them from affecting each other.

Timeout for metering GRPC calls set to 10 seconds.

This change also re-enables the separated application and quartz db connection pool feature (CB-8951 and CB-20455), which did not work because the quartz thread name was renamed.