CB-23151: Improve the parcel availability checker to handle parcel URL redirects
In this commit, I've added an extra config to the WebTarget to follow the redirects of the parcel URLs

The problem was that during the cluster upgrade preparation, we tried to verify the existence of the new Cloudera Manager rpm files then the request returned with an HTTP 301 status code.
With this improvement, the WebTarget will handle the redirects correctly.