CB-23260 Enable Azure storage account encryption by default
Azure Storage encryption is enabled for all storage accounts, including both Resource Manager and classic storage accounts.
Azure Storage encryption cannot be disabled. This recently started causing an issue where
Cloudbreak attempted to create a storage account to copy the VHD image to with disabled
encryption, but Azure responded with a 'FeatureNotSupportedForAccount' error.
Encryption was disabled by default, it was possible to enable it in the stack request through the encryptStorage field.
This field is now deprecated, CDP CLI does not set it.