CB-22528 Validate Azure Marketplace image terms at the beginning of a DL creation
Extended the Data Lake creation flow's validation step, with a request to CB
to validate the Azure Marketplace image terms on either the provided image or
on a selected iamge, which uses the same logic for selection as later
the CB stack creation flow.
Introduced a new endpoint to CB to select an image based on the provided inputs
and validate Azure marketplace image terms if applicable.