CB-23443 SecretToString and SecretProxy must be simplified
SecretToString had a very weird solution as it had two main usages, once as a Spring bean
and secondly as an Attribute Converter.
The latter one is used by Spring/Hibernate and in a later Hibernate version it is
constructed by a different class loader which cannot utilize the PostConstruct
initialization logic where the SecretService is initialized as static field.
Therefore, the passed SecretService "instance" became null causing all SecretProxy reads an NPE.

Co-authored-by: Dénes Berger <dberger@cloudera.com>
Co-authored-by: Richárd Doktorics <rdoktorics@cloudera.com>
Co-authored-by: Lajos Rodek <lrodek@cloudera.com>