CB-23304 Change instance type based on the metering instance checker job
Instance types are stored on hostgroup level, but the customer can change the instance type on individual instances manually on the provider. The MeteringInstanceCheckerJob will identify these instances and it will change the default instance type to the largest known instance type in the hostgroup (an instance is larger only if both the CPU and memory is larger than the default). This means all of the instances will be reported for meteringv2 with the largest instance type in the hostgroup no matter what the real instance type is.
The default instance type is stored in the template table and in launch configuration/launch template in case of AWS.