CB-23633 - Turning off geo-redundant backup globally
According to the agreement with bgaspar and khorvath I am turning off
geo-redundant backup.

It is very unlikely that customers would want to recover from a geo-replicated automatic PG backup, in case of a region loss. The only database instance that is of value is the SDX DB, for which we provide  backup to cloud storage. Customers can configure the underlying cloud storage to be geo-redundant, thus making the backup available across regions for DR purposes.
Furthermore, there is a fair chance that a low-level backup from the Data lake DB in geo1 cannot be simply restored to a Data lake DB in geo2, as e.g. the schema credentials will not match.
Customers should be encouraged to use SDX backup and restore which formally supports moving SDX metadata between environments