CB-23697 Enhance mock-infrastructure to support verification of recommendation Enhance mock-infrastructure to support verification of recommendation
As part of CB-23155, CB has been changed to reverify decommission candidates as part of downscale flow. For mock cloud provider, We are using mock-infra to return recommendation. Currently, it is alternatively returning upscale and downscale recommendation so that cluster is always scaled based on recommendation which is needed for load testing.

After this change, mock-infra will be called twice in case of downscale so current implementation will not return decommission candidates (since mock-infra will return recommendation for upscale). To solve this, We are going to make a change to support verification.

Below are the changes :

Additional query parameter actionType will be supported in mock-infra for YARN recommendation. If actionType = "Verify" then previous recommendation will be returned else recommendation will be toggled and returned.
Periscope will send query parameter actionType = "verify" as part of recommendation API for mock cloud provider.