CB-23695 Added separate handling for default blueprints
Default blueprints don't store secrets, so their JSONs can be stored in postgres and retrieved from there instead of getting them one-by-one from vault. The defaultblueprinttext column has been added to the db table, and this column stores the blueprint JSON in the case of default blueprints, and is null in the case of user managed blueprints. The blueprinttext column still contains the vault path for all blueprints, even the default ones. Using the Blueprint class's getBlueprintJsonText() method is recommended instead of the getBlueprintText() method, as it will return the blueprint JSON without calling the vault, if the Blueprint is a default blueprint.