CB-23772 Using database instancetype based on flexible server capability in the given region
For flexible server the "Standard_E4ds_v4" is not supported in all region, so cloudbreak will use the newest generation of instancetype
in the given region.
Datalake and Cloudbreak services won't send the instancetype to Redbeams service during flexible server creation or upgrade from single to flexible server.
In this case Redbeams will determine the newest version of the memory optimized instancetypes and will use it.
If the instancetype is given in the appropriate API call, Redbeams will use the given one.