CB-24042 Fallback to Azure Private Single Server if the delegated subnet id is not specified - Datalake
Scenario: customer has an existing environment with Private Single Server enabled. This means that the environment has 1 Private Dns Zone configured. If a new Data Lake gets created we would create a Public Flexible Server, but probably the customer does not want a public DB server. So instead of Flexible we should just fall back to Private Single Server as usual.

If customer specifies that the DB server should be Flexible, cloudbreak will reject the request to specify the delegated subnet id.