CB-23997 modify default root disk size to 200G, root disk migration at repair
In case of RHEL8 when you try to upgrade from 7.2.17 to 7.2.18 we don't have enough space for parcel download.
I set the default root disk size to 200G from 100G and we spin up new nodes at repair with 200G root disks if root disk is smaller than this value.