CB-24314 Add 'reverse-zone' flags to FreeIPA replica install command
We have seen in ENGESC-23363 that it can occur during replica install the reverse zone is created for a subnet which is already handled. Investigating the FreeIPA code it become clear it behaves a bit different when reverse zone is provided to the install command. Providing it should not cause any issue, but help the install process recognize that the right zone already exists avoiding unnecessary sub-zone creation.