CB-17625 Fix NPEs
UsageProto.CDPDatabaseDetails builder throws NPE for null values so fallbacks were added to empty strings
FreeIpaClientFactory.getFreeIpaClient throws an NPE but could not get to root cause from the logs, so extended logging with cause exceptions
ThreadBasedUserCrnProvider.getUserCrn() may return null in EventParameterFactory implementations, and the following Map.of calls failed with NPE. To solve this a fallback getUserCrnByResourceId was introduced to get the resource owner/creator. If that also does not return a value, an empty Map is returned. Another improvement was made to check if the gathered user crn still exists in UMS to avoid subsequent UMS NOT_FOUND failures.