CB-24524 Fix URI is `UNKNOWN` in Prometheus metrics
Spring 6.x seems to drop method name discovery:
https://github.com/spring-projects/spring-framework/wiki/Upgrading-to-Spring-Framework-6.x#parameter-name-retention

Based on the trial, the "-parameters" javac switch helps to solve
the issue where the "uri" part is not recognized in servlet context
even in Spring 6.
The details are obscure but the
DefaultServerRequestObservationConvention#uri cannot determine
the URI part because the context.pathPattern is null.
I presume this is a Spring (regression) issue.