CB-23695: Fixed defaultBlueprintText not getting saved to DB causing unnecessary Vault calls
The `isAddingDefaultBlueprintsNecessaryForTheUser` method uses the `mustUpdateTheExistingBlueprint` method, which didn't use to check if the defaultBlueprintText field has changed or not, therefore the `defaultblueprinttext` column in the DB stayed null, resulting in the `Blueprint.getBlueprintJsonText` method calling the Vault each time, causing an unnecessary slowdown.
The checking of the defaultBlueprintText field has been integrated into the `defaultBlueprintNotSameAsNewTexts` method's logic.