CB-24124: Added secret encryption to secrets in freeipa EC2 user data
Secret encryption for AWS is implemented by the `AwsEncryptionSdkCryptoConnector` class, which uses the AWS Encryption SDK with AWS KMS as the key provider to encrypt secrets.

Besides the input, we also give the encrypt function an encryption context. The encryption context is a key-value map placed beside every ciphertext as metadata. The values in the encryption context can be used to verify the integrity of the ciphertext on the cluster side. The AWS Encryption SDK makes it mandatory to be able to provide the same encryption context values during decryption to be able to retrieve the original secret. Currently we provide the `ENVIRONMENT_CRN` and the `SECRET_NAME` (which is the same name as the environment variable we will place the secret into) during encryption, inserting them into th encryption context.

The user data has been extended to feature the `SECRET_ENCRYPTION_ENABLED` flag and the `SECRET_ENCRYPTION_KEY_ARN` of the KMS key used, if secret encryption is enabled for the environment.

In the `UserDataBuilder` class, we determine if a value is a secret and needs to be encrypted, by checking if its key is in the `secretKeysAndNames` map, which is populated from the `secret-keys-and-names.json` file. This resource file is a single point, where we can collect the values we want to encrypt before we place them in the user data.