CB-24666 Check for CUSTOM cluster shape SDX create request
CDP CLI will now send CM/CDH GBNs inside of the internalStackV4Request so this can no longer be used for determining if a custom stack is requested, so instead the cluster shape is checked.
For this to work the runtime version has to be computed all the times to get the correct stack request from cdpConfigService. This also meant removing some tests around empty runtime use cases.