CB-23911 Add 1 min delay during redbeams startup on AWS
On AWS after DB is started according to the provider, sometimes there is a restart of the instance/server.
This makes redbeams believe the DB is started and reports back to the services cluster can be started.
After starting the instance there is another state check. By delaying this check with one min, we gica AWS enough time to transition the database state to `rebooting`.
This is just a workaround, as CM should fail on such scenario, and systemd would restart it which would mitigate such issues transparently.