CB-24813: Ensured no calls are made to the Vault when loading default blueprints during first run
The `blueprintText` field is no longer filled with any Vault path in the case of default blueprints. Instead in the application it is a `Secret.EMPTY` value, and a new converter was added (`BlueprintSecretToString`), which converts the Secret object to the String `"BLANK"` if the blueprinttext is a `Secret.EMPTY` and vice versa.

The value `Secret.EMPTY` was chosen, because both `raw` and `secret` fields of the `Secret` object need to be `null` for the secret saving and deleting aspect code not to throw errors without modifying them.

The value `"BLANK"` was chosen for in-DB value, as the `blueprinttext` field cannot be null.