CB-24805 Quartz requires a blocking threadpool/executor implementation
Quartz uses SimpleThreadPool by default which blocks until it can provide a worker thread for the scheduler. If we configure a custom executor for Quarz, Spring uses the LocalTaskExecutorThreadPool class to configure a custom SchedulerFactoryBean and map the required ThreadPool interface calls to the custom executor. Unfortunately, this class was implemented by Spring in a non-blocking way which is incompatible with the way Quartz works.

Created a custom ThreadPool implementation to configure a custom TimedSimpleThreadPoolTaskExecutor for Quartz and delegate calls from the Quartz scheduler to the underlying thread pool.
 This class is loaded and initalized during Quartz scheduler creation and responsible to load the related task executor bean from Spring.
 The TimedSimpleThreadPoolTaskExecutor class is an enhanced SimpleThreadPool implementation to provide the required metrics (threadpool size, usage, etc.).