CB-24877 Update max hostname length config on FreeIPA when upgrading to RHEL8
The current implementation sets up the max hostname length only during provisioning.
Adjusted the `FreeIpaPostInstallService` to check the config and update it every time it's called. Thus this will run after every upscale too.
The logic is adopted, that if the client is not connected to a RHEL8 freeipa instance, then look for one if available and connect to it for setting the config. The config can not be set on a CentOS7 instance, but it can be read if there is a RHEL8 already available.
By default the client is created against the primary gateway which is replaced last during an upgrade, this is the reason why we needed the above workaround.
If no RHEL8 instance is available for some magical reason (it got deleted on provider) then the logic will ignore this silently for now. An upscale/repair/upgrade would update the config.