CB-23825 Implement creating new instance with the right FQDN
Implemented the required flow steps to create the new instance and install the new FreeIPA on it with the same FQDN and private ID as in the request.
A new flow step is added to finalize the instance termination which was missing.
Also stack status updates added to the flow steps to ensure the right state.