CB-24901 Targeting Subnets for Load Balancers not working
In case of Azure, DataHub the endpoint gateway subnet setting
was not propagated in the Azure Network Parameters thus
the ARM template could not be properly filled and was causing
Deployment errors. This is now fixed.
Note: the `endpointGatewaySubnetId` field in the
`AzureNetworkV4Parameters` class is a hidden field from the
OpenAPI's point of view.