CB-23826 Implement backup validation step for FreeIPA rebuild
Added states and script to download the backup from cloudstorage and validate the files are present and the header
contains the right values.
The base for the script was the backup script.

Modified FreeIPA client creation to make it possible to create a client if the state is unreachable furing post install.
It's required to make move forward possible when retry is required but somehow the instance status is stuck in wrong state.