CB-25044 Introducing virtual threads to cloudbreak services.
Virtual thread based thread pools:

Almos all thread pools are switched to use virtual threads if the
spring.threads.virtual.enabled flag is true. The below thread pools are
switched to virtual threads:

- Tomcat thread pools,
- Flow handler thread pools,
- Quartz thread pools,
- Azure client thread pool,
- user sync realted thread pools,
- other utility thread pools such as cloudbreakListeningScheduledExecutorService.

Replaced synchronized with lock:

Virtual threads can't give back control to the jvm virtual thread
scheduler if synchronized is used. A ReentrantLock is used instead where
possible.

Fixed and introduced monitoring:

A couple of thread pools were not monitored or not monitored in the
proper way. For example execution times of the flow related tasks were
not monitored properly because we did not use the TimedExecutorService
returned by ExecutorServiceMetrics as a bean. I also inrtoduced
monitoring for ForkJoinPool.commonPool() and for the VirtualThread
thread pool which is also a ForkJoinPool. Monitoring the virtual thread
handling ForkJoinPool is a bit dirty but I couldn't find another way
than accessing the VirtualThread.scheduler field via reflection. For
this the --add-opens java.base/java.lang=ALL-UNNAMED jvm option is
needed. Without that the application can be started but the underlying
virtual thread scheduler won't be monitored.

Prometheus in integration tests:

To collect metrics about cloudbreak services I introduced prometheus as
a docker container during integrations tests. It is pointed to
cloudbreak services and the metrics are compressed into the
prometheus-data.tar.gz file. Spinning up a local prometheus and pointing
it to it as a data folder the metrics can be viewed.

Increased integration test thread count:

With virtual threads the services consume less memory therefore the
TestNG thread count was increased to 16 when mock provider is used.

Fixed flaky integration tests:

RequestResponseStorageService implementation was wrong and it returned
the first records if the call recording was enabled not taking into
account the crn of the entity. This could come from totally unrelated
tests and it was a miracle that it worked. Now mock infrastructure calls
are recorded every time and are deleted at the end of every test.

Tested cases:

- Env, Data Lake, Data Hub creation, stop, start
- Data Lake repair, upgrade
- Data Lake repair, upscale, downscale, node deletion, upgrade
- User sync cancellation by timeout
- Env, Data Lake, Data Hub deletion
- Checked the data source routing and azure client retry based on thread name
- We don't store large objects in thread locals, therefore it shouldn't
be a problem