CB-25226 FreeIPA upgrade should offer RHEL8 as an option when upgrading from CentOS7
When collecting possible target images for UI and cli which is shown to the user we ensure we add RHEL8,
even if the RHEL8 image is older then CentOS7.
This implemented by depending the defaultos configured for the application. So if RHEL9 is introduced and set as default,
then it would work too.
OsType ordinal is used to decide which release is the latest (higher is newer). This validated by a unit test to avoid forgetting to update.