CB-25369 Reorder OS upgrade flowchain to bring embedded upgrade before VM replacement
Instead of upgrading the database after the new instances created, it's done before the replacement.
It should not cause any issue if the archive is available.
Upgrade is forbidden only if no CDP_POSTGRES_UPGRADE_EMBEDDED is provided and
current DB version is 10. At this point unfortunately we can't filter for target OS version in the flow
as it still shows CentOS7 to ensure the right parcels are downloaded.