CB-25414 Use load balancing for WEBHBASE Knox service
Change HA parameters in the Knox topology templates so that the WEBHBASE
(aka HBase REST server) service is using sticky load balancing.

This is required both for working load balancing, and for being able to use
the Knox session cookies for authentication without overloading the identity infra
by accessing it for every request.