CB-25403 Refactoring the StaticApplicationContext and moving out the config properties to HibernateCircuitBreakerConfigProvider.
The reason why it is moved out because we can fetch the spring config values in advance in startup time, therefore it is going to be more efficient during operation, since we don't need to call spring. The main reason why we would like to avoid calling spring is that org.springframework.core.env.PropertiesPropertySource.getPropertyNames(PropertiesPropertySource.java:50) contains a synchronized block.