CB-23602 Remove healthcheck from cluster proxy registration
As FMS doesn't depend on CP load balancing any more, we could remove the health check on CP side.
Also this renders the load balanced endpoint registration useless, as CP would not be able to determine which
server to connect to, and would connect randomly.
Tight now FMS does it's own health check and chooses the destination based on this information.