CB-25293 Implement YarnInstanceConnector.check
With this change the conclusion checker, syncer, and other high level implementations will be able to query YARN instance statuses. This may help debug issues when the YARN instances have issues.
Implemented in a way to reuse existing implementation from YarnMetadataCollector.collect as it retrieves YARN application details. This needed the application name stored in the CloudInstance which required quite a refactor to get into the StackToCloudStackConverter.buildCloudInstanceParameters method.