CB-25515 Enable and trigger FreeIPA SID generation
With the new version of FreeIPA it's required for users to have SID configured.
When upgrading from CentOS7 to latest Rhel8, the users already existing on CentOS7 doesn't have this attribute.
In this change we call this FreeIPA command after every upscale if a RHEL8 instance found.
In case of a retryable error, the call is retried, otherwise for now the error is just logged, but the upscale is allowed.
Unfortunately there is no way to check easily if this config turned on, or if every user has the property present.
Only method would be to query all users and check, but it actually would take more time and resources then the call itself.