CB-25569 Do not fetch unavailable env for FreeIPA usersync
As mow dev is flooded with delete failed envs due to CDL bug, it's causing memory issues,
when FMS is trying to fetch all the env in the account for partial user sync.
This change filters freeipa stack fetching based on it's availability status,
so those won't be fetched for usersync which are not usable anyway.