CB-25666 Do not run SID generation after every upscale for RHEL8
SID generation should run only once, when upgrding to RHEL8 and the first RHEL8 instance added.
When the FreeIPA is provisioned with RHEL8, this is not required.

This is required as this command triggars and LDAP/SLAPD/DIRSRV restart, which sometimes hangs
until the 10min systemd timeout happens and force kills the service before restarting.