CB-25725 Fetch accelerated network enabled instance dynamically
Instead of depending on a file, accelerated network support is fetched dynamically similarly to host encryption.
To support this feature, capability fetching is a bit generalized on Azure side.

During template creation now these values are fetched from Azure. To support this a wrapper method is introduced, so it works
with CloudCredential instead of the Extended one, as it's not required anyway. The 2 methods depends on the same cache
as parameters and return value otherwise are the same.