CB-25777 Add embedded DB upgrade after runtime upgrade for datahub
As it's possible to run runtime upgrade with RHEL8 target, it will switch the stack image to RHEL8 while not running OS upgrade (this is on purpose for DH). This means while all the instances are on CentOS7, the stack image and all autoscaling group is set to RHEL8 image.
If the CentOS7 is running on postgres 10, then a repair would fail for the master instance as RHEL8 doesn't have postgres 10 installed, and as there was no OS upgrade, there was no embedded upgrade.
To avoid such issues, we'll add embedded DB upgrade flow steps after DH runtime upgrade, so if a repair would be required it won't fail.