CB-25786: Wait for restart operation to finish before service starts in case of patch upgrade
In this commit I've added an extra polling for an existing restart command in the CM before we try to start the services after parcel activation.
We need this extra check because if the flow restarts after the restart command the start command will fail because there is an existing restart command.