CB-25670 Enable Medium Duty HA datalake runtime upgrade with entitlement
The goal of this change, to have an option where we can enable a customer to upgrade their datalake to 7.2.18 runtime and above with the MEDIUM_DUTY_HA datalake shape.
As a result we are not showing this option to the customer when they list the available candidates. But when a customer file a case to enable that, we grant the entitlement, and the list should contain all the runtime version (image) whicn available on that setup.
The current implementation does not support any easy way to list all the images the customer. With the entitlement exetension that should be an easy turn on button on the SDX upgrade. Also enable the medium duty creation with higher runtime
Therefore this modification contains the following:
1) Add the new Entitlement to the Entitlement.java class. (The new Entitlement is `MEDIUM_DUTY_UPGRADE_ON_HIGHER_RUNTIME`)
2) Add the `isSdxRuntimeUpgradeEnabledOnMediumDuty` method to the EntitlementService. Which evaluate the current state of the entitlement in the Account by account id.
3) Modification in the current `SdxUpgradeFilter` class `filterImageByShape` method. The modification is added a new condition to the if branch, which will be true, if the entitlement is granted on the account.
Justification: The Porche who is our customer requesting that change.

The change has some refactor on the SdxRuntimeUpgradeSericeTest which remove the current code duplication and extract them into two method.
1) which setup the required IAM mock for the testing
2) which setup the ImageV4Response with the required runtimes