CB-25400 External DB upgrade should fail if we can't download the PostgreSQL packages
As part of the DB upgrade process we will try to download the packages from archive.cloudera.com and if the download fails we should throw an error and the upgrade should not be continued. If the customer manually installs the correct packages we should be able to detect that and continue with the upgrade.