CDPSDX-6555  fix restore validation param
Fix the Datalake Resize restore validation parameter from the hardcoded FALSE to the dynamic variable. The goal of this change is to execute the restore validation without executing the real database restore. This change will prevent some unexpected rollback in the ranger or in the Hive database.

Executed tests:

Reszie Validation from LD to EDL (ranger Policy check success, the validation does not rollbacked the changes)
Resize Validation from MD to EDL (ranger Policy check success, the validation does not rollbacked the changes)