CB-25907 Change resource name suffix from datetime to epoch millis
As E2E tests shown that there is a slight chance of resource name collision
in case the tests are run parallel and the resource name is generated at
the same second. This could be a problem at customer side as well, becuase
environments/clusters might get created via an automatism, resulting in a
a similar error.

The chances are now divided by 1,000 by utilizing the epoch millis, that is,
the resource names only collide if they are generated at the same millisecond.

Therefore the constant "hash" lenght is set to 13 (the current length of
epoch millis, which is good for a good 200 years).

Adjusted unit tests accordingly.