B-26015 SdxResize recipe handling fix
The goal of this modificaiton that the resize workflow handle the attached recipes of the original datalake.
Before this change the resize workflow was not able to copy the attached recipes, therefore, all the resized datalake do not have any recipes on thir hosts.
The logic which is appended to handle the current recipes under the resize, is to get the Stack from the original datalake and gather the recipes and hostgroups, and reattach them to the new stackrequest which will created under the resize.
The change tested locally, with cbd. Tests contains the regression tests which is aplain resize without any recipes. The resize was successful. The new action test is that the attach multiple recipes to multiple hosts and resize the datalake. The resized datalake contains the same recipes on the same hosts. So we can say the test was successful.

Extension: Little modification on the sdxSerivceTest to remove the hardcoded random values and use a declared class field. This make the understanding of the unit tests easier