CB-24356 Add Knox GW / IDBroker keystores in secret location if secret encryption is enabled for Environment See OPSAPS-69418. If secret encryption is enabled then Knox GW/IDBroker keystores should be stored in secret location. This location is same as secret location configured as part of provisioning the cluster.
Following needs to be done if secret encryption is enabled:

Add pillar for knoxGatewaySecurityDir and knoxIdBrokerSecurityDir. These values as same as configured for gateway_security_dir and idbroker_security_dir as part of provisioning.
Keystores should be placed at {knoxGatewaySecurityDir}/keystores and {knoxIdBrokerSecurityDir}/keystores.
owner and group should set as knox:knox.
mode should be set as 640.